+++ src/processing.h
@@ -8,7 +8,7 @@
 #ifndef SRC_PROCESSING_H_
 #define SRC_PROCESSING_H_
 
-double process_channel(double *c2, double *psd_ch_sample, double *correction_base, double q_ch, int64_t m_ch);
+double process_channel(double *c2, double *psd_ch_sample, double *correction_base, double q_ch, double m_ch);
 
 
 #endif /* SRC_PROCESSING_H_ */
